\name{rpgm-package}
\alias{rpgm-package}
\alias{rpgm}
\docType{package}
\title{
rpgm
}
\description{
Fast Simulation of Normal Random Variables
}
\details{

Ziggurat method in order to simulate normal random variables approximately four times faster than the usual rnorm(), reference : MARSAGLIA, George, TSANG, Wai Wan, et al. The ziggurat method for generating random variables. Journal of statistical software, 2000, vol. 5, no 8, p. 1-7.

}
\author{
Nicolas Baradel	

Maintainer: Nicolas Baradel - PGM Solutions
}
\references{
\href{http://pgm-solutions.com/packages}{http://pgm-solutions.com/packages}
}


\keyword{ package }
%\seealso{

%}
\examples{
	rpgm.rnorm(5)
}
